powerapi/requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.1.1
daphne==2.5.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==3.0.9
djangorestframework==3.12.1
hiredis==1.1.0
html5lib==1.0.1
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
iniconfig==1.0.1
ipaddr==2.2.0
llvmlite==0.34.0
lockfile==0.12.2
msgpack==1.0.0
networkx==2.5
numba==0.51.2
numpy==1.19.2
packaging==20.3
pandapower==2.4.0
pandas==1.1.3
pep517==0.8.2
pluggy==0.13.1
progress==1.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytest==6.1.1
pytest-asyncio==0.14.0
pytest-django==3.10.0
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
scipy==1.5.2
service-identity==18.1.0
six==1.14.0
sqlparse==0.4.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
webencodings==0.5.1
xlrd==1.2.0
XlsxWriter==1.3.6
zope.interface==5.1.2
